add skip-file option, for keeping dependencies

+++ Build/clean-deps.pl
@@ -13,11 +13,13 @@ use strict;
 use Getopt::Long;
 
 my $depfile;
+my $skipfile;
 my $cmsdist;
 my $remove = 0;
-GetOptions("dep-file=s" => \$depfile,
-	   "cmsdist=s"  => \$cmsdist,
-	   "remove"     => \$remove);
+GetOptions("dep-file=s"  => \$depfile,
+	   "skip-file=s" => \$skipfile,
+	   "cmsdist=s"   => \$cmsdist,
+	   "remove"      => \$remove);
 
 die "check arguments" unless $depfile && $cmsdist;
 
@@ -30,13 +32,26 @@ while (<DEPS>) {
 }
 close DEPS or die $!;
 
-print join("\n", @deps), "\n";
+my @skip;
+if ($skipfile) {
+    open SKIP, '<', $skipfile or die $!;
+    while (<SKIP>) {
+	chomp;
+	next unless $_;
+	push @skip, $_;
+    }
+    close SKIP or die $!;
+}
 
 foreach my $file ( <$cmsdist/*> ) {
+    next unless -f $file;
     my @matches = grep($file =~ /$_/, @deps);
     my $is_dep = @matches ? 1 : 0;
-    if ($is_dep) {
+    my $skip = grep($file =~ /$_/, @skip) ? 1 : 0;
+    if ($is_dep && !$skip) {
 	print "$file is a dependency (", join(' ', @matches), ")... keep\n";
+    } elsif ( $skip ) {
+	print "$file on skip list... keep\n";
     } else {
 	print "$file is not a dependency... remove\n";
 	unlink $file if $remove;
